backend/pyproject.toml
[project]
name = "learnsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.2",
    "email-validator>=2.3.0",
    "fastapi>=0.120.0",
    "firecrawl>=4.5.0",
    "jwt>=1.4.0",
    "motor>=3.7.1",
    "openai>=2.6.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "twilio>=9.8.4",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
